+++ cpp/src/arrow/CMakeLists.txt
@@ -644,6 +644,9 @@ endif()
 if(ARROW_WITH_OPENTELEMETRY)
   list(APPEND ARROW_UTIL_SRCS util/tracing_internal.cc)
 endif()
+if(ARROW_WITH_SIMDJSON)
+  list(APPEND ARROW_UTIL_SRCS util/json_writer_internal.cc util/simdjson_internal.cc)
+endif()
 if(ARROW_WITH_SNAPPY)
   list(APPEND ARROW_UTIL_SRCS util/compression_snappy.cc)
 endif()
@@ -663,6 +666,12 @@ endforeach()
 foreach(ARROW_UTIL_TARGET ${ARROW_UTIL_TARGETS})
   target_link_libraries(${ARROW_UTIL_TARGET} PRIVATE ${ARROW_XSIMD})
 endforeach()
+
+if(ARROW_WITH_SIMDJSON)
+  foreach(ARROW_UTIL_TARGET ${ARROW_UTIL_TARGETS})
+    target_link_libraries(${ARROW_UTIL_TARGET} PRIVATE arrow::simdjson)
+  endforeach()
+endif()
 
 if(ARROW_USE_BOOST)
   foreach(ARROW_UTIL_TARGET ${ARROW_UTIL_TARGETS})
@@ -1088,7 +1097,6 @@ if(ARROW_JSON)
                            json/chunker.cc
                            json/converter.cc
                            json/from_string.cc
-                           json/json_writer_internal.cc
                            json/object_parser.cc
                            json/parser.cc
                            json/reader.cc)
