+++ cpp/src/arrow/extension/fixed_shape_tensor.cc
@@ -27,8 +27,8 @@
 
 #include "arrow/array/array_nested.h"
 #include "arrow/array/array_primitive.h"
-#include "arrow/json/json_writer_internal.h"
 #include "arrow/tensor.h"
+#include "arrow/util/json_writer_internal.h"
 #include "arrow/util/logging_internal.h"
 #include "arrow/util/print_internal.h"
 #include "arrow/util/simdjson_internal.h"
