+++ cpp/src/arrow/json/meson.build
@@ -22,7 +22,6 @@ exc = executable(
         'chunker_test.cc',
         'converter_test.cc',
         'from_string_test.cc',
-        'json_writer_internal_test.cc',
         'parser_test.cc',
         'reader_test.cc',
     ],
