+++ cpp/src/arrow/util/json_writer_internal.h
