+++ cpp/src/parquet/encryption/local_wrap_kms_client.cc
@@ -17,8 +17,8 @@
 
 #include <string_view>
 
-#include "arrow/json/json_writer_internal.h"
 #include "arrow/json/object_parser.h"
+#include "arrow/util/json_writer_internal.h"
 #include "arrow/util/secure_string.h"
 
 #include "parquet/encryption/key_toolkit_internal.h"
